@@ -1,8 +1,8 @@
 package com.gorillamoa.routines.fragment
 
 
-import android.arch.lifecycle.ViewModelProviders
-import android.support.v4.app.Fragment
+import androidx.lifecycle.ViewModelProviders
+import androidx.fragment.app.Fragment
 import com.gorillamoa.routines.viewmodel.TaskViewModel
 
 class LiveDataHeadless: Fragment(){
